+++ qtgui/plotter.h
@@ -42,8 +42,8 @@ public:
         resizeEvent(NULL);
     }
 
-    void setNewFttData(double *fftData, int size);
-    void setNewFttData(double *fftData, double *wfData, int size);
+    void setNewFttData(float *fftData, int size);
+    void setNewFttData(float *fftData, float *wfData, int size);
 
     void setCenterFreq(quint64 f);
     void setFreqUnits(qint32 unit) { m_FreqUnits = unit; }
@@ -85,9 +85,9 @@ public:
         drawOverlay();
     }
 
-    void setMaxDB(double max);
-    void setMinDB(double min);
-    void setMinMaxDB(double min, double max);
+    void setMaxDB(float max);
+    void setMinDB(float min);
+    void setMinMaxDB(float min, float max);
 
     void setFontSize(int points) { m_FontSize = points; }
     void setHdivDelta(int delta) { m_HdivDelta = delta; }
@@ -96,7 +96,7 @@ public:
     void setFreqDigits(int digits) { m_FreqDigits = digits>=0 ? digits : 0; }
 
     /* Determines full bandwidth. */
-    void setSampleRate(double rate)
+    void setSampleRate(float rate)
     {
         if (rate > 0.0)
         {
@@ -105,7 +105,7 @@ public:
         }
     }
 
-    double getSampleRate(void)
+    float getSampleRate(void)
     {
         return m_SampleFreq;
     }
@@ -135,7 +135,7 @@ public slots:
     void setFftPlotColor(const QColor color);
     void setFftFill(bool enabled);
     void setPeakHold(bool enabled);
-    void setPeakDetection(bool enabled, double c);
+    void setPeakDetection(bool enabled, float c);
     void updateOverlay();
 
 protected:
@@ -163,87 +163,91 @@ private:
     int xFromFreq(qint64 freq);
     qint64 freqFromX(int x);
     qint64 roundFreq(qint64 freq, int resolution);
-    bool isPointCloseTo(int x, int xr, int delta){return ((x > (xr-delta) ) && ( x<(xr+delta)) );}
+    bool isPointCloseTo(int x, int xr, int delta)
+    {
+        return ((x > (xr - delta)) && (x < (xr + delta)));
+    }
+
     void clampDemodParameters();
 
     void getScreenIntegerFFTData(qint32 plotHeight, qint32 plotWidth,
-                                 double maxdB, double mindB,
+                                 float maxdB, float mindB,
                                  qint64 startFreq, qint64 stopFreq,
-                                 double *inBuf, qint32 *outBuf,
+                                 float *inBuf, qint32 *outBuf,
                                  qint32 *maxbin, qint32 *minbin);
 
     void zoomStepX(float factor, int x);
 
-    bool m_PeakHoldActive;
-    bool m_PeakHoldValid;
-    qint32 m_fftbuf[MAX_SCREENSIZE];
-    qint32 m_fftPeakHoldBuf[MAX_SCREENSIZE];
-    double *m_fftData;     /*! pointer to incoming FFT data */
-    double *m_wfData;
-    int     m_fftDataSize;
+    bool        m_PeakHoldActive;
+    bool        m_PeakHoldValid;
+    qint32      m_fftbuf[MAX_SCREENSIZE];
+    qint32      m_fftPeakHoldBuf[MAX_SCREENSIZE];
+    float      *m_fftData;     /*! pointer to incoming FFT data */
+    float      *m_wfData;
+    int         m_fftDataSize;
 
-    int m_XAxisYCenter;
-    int m_YAxisWidth;
+    int         m_XAxisYCenter;
+    int         m_YAxisWidth;
 
-    eCapturetype m_CursorCaptured;
-    QPixmap m_2DPixmap;
-    QPixmap m_OverlayPixmap;
-    QPixmap m_WaterfallPixmap;
-    QColor m_ColorTbl[256];
-    QSize m_Size;
-    QString m_Str;
-    QString m_HDivText[HORZ_DIVS_MAX+1];
-    bool m_Running;
-    bool m_DrawOverlay;
-    qint64 m_CenterFreq;
-    qint64 m_FftCenter;
-    qint64 m_DemodCenterFreq;
-    bool m_CenterLineEnabled;  /*!< Distinguish center line. */
-    bool m_FilterBoxEnabled;   /*!< Draw filter box. */
-    bool m_TooltipsEnabled;     /*!< Tooltips enabled */
-    int m_DemodHiCutFreq;
-    int m_DemodLowCutFreq;
-    int m_DemodFreqX;		//screen coordinate x position
-    int m_DemodHiCutFreqX;	//screen coordinate x position
-    int m_DemodLowCutFreqX;	//screen coordinate x position
-    int m_CursorCaptureDelta;
-    int m_GrabPosition;
-    int m_Percent2DScreen;
+    eCapturetype    m_CursorCaptured;
+    QPixmap     m_2DPixmap;
+    QPixmap     m_OverlayPixmap;
+    QPixmap     m_WaterfallPixmap;
+    QColor      m_ColorTbl[256];
+    QSize       m_Size;
+    QString     m_Str;
+    QString     m_HDivText[HORZ_DIVS_MAX+1];
+    bool        m_Running;
+    bool        m_DrawOverlay;
+    qint64      m_CenterFreq;
+    qint64      m_FftCenter;
+    qint64      m_DemodCenterFreq;
+    bool        m_CenterLineEnabled;  /*!< Distinguish center line. */
+    bool        m_FilterBoxEnabled;   /*!< Draw filter box. */
+    bool        m_TooltipsEnabled;     /*!< Tooltips enabled */
+    int         m_DemodHiCutFreq;
+    int         m_DemodLowCutFreq;
+    int         m_DemodFreqX;		//screen coordinate x position
+    int         m_DemodHiCutFreqX;	//screen coordinate x position
+    int         m_DemodLowCutFreqX;	//screen coordinate x position
+    int         m_CursorCaptureDelta;
+    int         m_GrabPosition;
+    int         m_Percent2DScreen;
 
-    int m_FLowCmin;
-    int m_FLowCmax;
-    int m_FHiCmin;
-    int m_FHiCmax;
-    bool m_symetric;
+    int         m_FLowCmin;
+    int         m_FLowCmax;
+    int         m_FHiCmin;
+    int         m_FHiCmax;
+    bool        m_symetric;
 
-    int    m_HorDivs;   /*!< Current number of horizontal divisions. Calculated from width. */
-    int    m_VerDivs;   /*!< Current number of vertical divisions. Calculated from height. */
-    double m_MaxdB;
-    double m_MindB;
-    qint32 m_dBStepSize;
-    qint32 m_Span;
-    double m_SampleFreq;    /*!< Sample rate. */
-    qint32 m_FreqUnits;
-    int m_ClickResolution;
-    int m_FilterClickResolution;
+    int         m_HorDivs;   /*!< Current number of horizontal divisions. Calculated from width. */
+    int         m_VerDivs;   /*!< Current number of vertical divisions. Calculated from height. */
+    float       m_MaxdB;
+    float       m_MindB;
+    qint32      m_dBStepSize;
+    qint32      m_Span;
+    float       m_SampleFreq;    /*!< Sample rate. */
+    qint32      m_FreqUnits;
+    int         m_ClickResolution;
+    int         m_FilterClickResolution;
 
-    int m_Xzero;
-    int m_Yzero;  /*!< Used to measure mouse drag direction. */
-    int m_FreqDigits;  /*!< Number of decimal digits in frequency strings. */
+    int         m_Xzero;
+    int         m_Yzero;  /*!< Used to measure mouse drag direction. */
+    int         m_FreqDigits;  /*!< Number of decimal digits in frequency strings. */
 
-    int m_FontSize;  /*!< Font size in points. */
-    int m_HdivDelta; /*!< Minimum distance in pixels between two horizontal grid lines (vertical division). */
-    int m_VdivDelta; /*!< Minimum distance in pixels between two vertical grid lines (horizontal division). */
+    int         m_FontSize;  /*!< Font size in points. */
+    int         m_HdivDelta; /*!< Minimum distance in pixels between two horizontal grid lines (vertical division). */
+    int         m_VdivDelta; /*!< Minimum distance in pixels between two vertical grid lines (horizontal division). */
 
-    quint32 m_LastSampleRate;
+    quint32     m_LastSampleRate;
 
-    QColor m_FftColor, m_FftCol0, m_FftCol1, m_PeakHoldColor;
-    bool m_FftFill;
+    QColor      m_FftColor, m_FftCol0, m_FftCol1, m_PeakHoldColor;
+    bool        m_FftFill;
 
-    double m_PeakDetection;
-    QMap<int,int> m_Peaks;
+    float       m_PeakDetection;
+    QMap<int,int>   m_Peaks;
 
-    QList< QPair<QRect, qint64> > m_BookmarkTags;
+    QList< QPair<QRect, qint64> >     m_BookmarkTags;
 };
 
 #endif // PLOTTER_H
